include token/jwk package from auth
also, reorganize `rest`, `http` and `httpsupport`, moving
all possible content in `httpsupport` to avoid the name
confusion/clash with `net/http`

fixes #18
