+++ httpsupport/http_client_option.go
@@ -1,4 +1,4 @@
-package rest
+package httpsupport
 
 import (
 	"net/http"
